some bug fix for slam6d and fbr
git-svn-id: svn://svn.code.sf.net/p/slam6d/code/trunk@840 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/slam6d/gapx6D.cc
@@ -427,7 +427,7 @@ double gapx6D::doGraphSlam6D(Graph gr, vector <Scan *> allScans, int nrIt)
 
       double dummy_sum;
       Scan::getPtPairs(ptpairs[i], FirstScan, SecondScan, thread_num,
-                       (int)my_icp->get_rnd(), (int)max_dist_match2_LUM, dummy_sum,
+                       (int)my_icp->get_rnd(), max_dist_match2_LUM, dummy_sum,
                        centroids_m[i], centroids_d[i]);
 
       // faulty network
